+++ build.sh
@@ -2,7 +2,9 @@
 set -eu
 
 # add configuration
-. ./config/env.docker.sh
+source ./config/env.docker.sh
+source ./artifact-lang.sh
 
 export NODE_OPTIONS='--max-old-space-size=4096'
+retrieveLanguageParameters
 npm run build