+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Blip is the web front end for YourLoops system.
 It is based on Tidepool Blip 1.27.
 
 ## Unreleased
+### Added
+- YLP-247 Add Dutch language
 ### Changed
 - YLP-206 allow name edition for patient profile
 - YLP-203 Change menu item "System settings" to "DBL settings"